Restructure web3-eth-abi interface
Originally, it had been my impression that the `types` input in
`encodeCallSignature` was unnecesary and redundant [1] and then when I
needed to expose `abi.encodeFunctionSignature`, I took it as a chance to
address this problem.

However, it turns out that `web3-abi-eth`'s `encodeFunctionCall` needs
an input variable's name and I think it's because of Solidity tuples.
At first, I tried enabling web3-eth-abi to take a string as a valid
functionName input for `encodeFunctionCall` - but then found the tuple
issue mentioned above [2].

So I came to the conclusion that it is probably best to simply expose
as much of the `web3-eth-abi` interface as we currently need and link to
their docs. It's a rather big breaking change but also direly necessary.

References:

- 1: https://github.com/rugpullindex/eth-fun/issues/15
- 2: https://github.com/ChainSafe/web3.js/pull/5229